+++ lib/src/display/display_object.dart
@@ -5,7 +5,11 @@ import 'package:flutter/painting.dart' as painting;
 import '../../graphx.dart';
 
 abstract class GDisplayObject
-    with DisplayListSignalsMixin, RenderSignalMixin, MouseSignalsMixin, DisplayMasking {
+    with
+        DisplayListSignalsMixin,
+        RenderSignalMixin,
+        MouseSignalsMixin,
+        DisplayMasking {
   GDisplayObjectContainer? $parent;
   static GDisplayObject? $currentDrag;
   static GRect? $currentDragBounds;
@@ -123,11 +127,9 @@ abstract class GDisplayObject
     }
   }
 
-  void $dispatchMouseCallback(
-    MouseInputType type,
-    GDisplayObject object,
-    MouseInputData input,
-  ) {
+  void $dispatchMouseCallback(MouseInputType type,
+      GDisplayObject object,
+      MouseInputData input,) {
     if (mouseEnabled) {
       var mouseInput = input.clone(this, object, type);
       switch (type) {
@@ -149,7 +151,8 @@ abstract class GDisplayObject
           $onMouseMove?.dispatch(mouseInput);
           break;
         case MouseInputType.up:
-          if ($mouseDownObj == object && ($onMouseClick != null || $onMouseDoubleClick != null)) {
+          if ($mouseDownObj == object &&
+              ($onMouseClick != null || $onMouseDoubleClick != null)) {
             var mouseClickInput = input.clone(this, object, MouseInputType.up);
             $onMouseClick?.dispatch(mouseClickInput);
 
@@ -200,7 +203,12 @@ abstract class GDisplayObject
 
   double get mouseX {
     if (inStage) {
-      return globalToLocal(_sHelperPoint.setTo(stage!.pointer!.mouseX, 0)).x;
+      return globalToLocal(
+        _sHelperPoint.setTo(
+          stage!.pointer!.mouseX,
+          stage!.pointer!.mouseY,
+        ),
+        _sHelperPoint,).x;
     } else {
       throw 'To get mouseX object needs to be a descendant of Stage.';
     }
@@ -208,7 +216,12 @@ abstract class GDisplayObject
 
   double get mouseY {
     if (inStage) {
-      return globalToLocal(_sHelperPoint.setTo(0, stage!.pointer!.mouseY)).y;
+      return globalToLocal(
+        _sHelperPoint.setTo(
+          stage!.pointer!.mouseX,
+          stage!.pointer!.mouseY,
+        ),
+        _sHelperPoint,).y;
     } else {
       throw 'To get mouseY object needs to be a descendant of Stage.';
     }
@@ -228,10 +241,18 @@ abstract class GDisplayObject
   Object? userData;
   String? name;
 
-  double _x = 0, _y = 0, _scaleX = 1, _scaleY = 1, _rotation = 0;
-  double _pivotX = 0, _pivotY = 0;
-  double _skewX = 0, _skewY = 0;
-  double _z = 0, _rotationX = 0, _rotationY = 0;
+  double _x = 0,
+      _y = 0,
+      _scaleX = 1,
+      _scaleY = 1,
+      _rotation = 0;
+  double _pivotX = 0,
+      _pivotY = 0;
+  double _skewX = 0,
+      _skewY = 0;
+  double _z = 0,
+      _rotationX = 0,
+      _rotationY = 0;
 
   double get rotationX => _rotationX;
 
@@ -592,18 +613,16 @@ abstract class GDisplayObject
     }
   }
 
-  void $updateTransformationMatrices(
-    double? x,
-    double? y,
-    double pivotX,
-    double pivotY,
-    double scaleX,
-    double scaleY,
-    double skewX,
-    double skewY,
-    double rotation,
-    GMatrix out,
-  ) {
+  void $updateTransformationMatrices(double? x,
+      double? y,
+      double pivotX,
+      double pivotY,
+      double scaleX,
+      double scaleY,
+      double skewX,
+      double skewY,
+      double rotation,
+      GMatrix out,) {
     out.identity();
     if (skewX == 0 && skewY == 0) {
       /// optimization, no skewing.
@@ -692,7 +711,8 @@ abstract class GDisplayObject
     return out;
   }
 
-  static GDisplayObject _findCommonParent(GDisplayObject obj1, GDisplayObject obj2) {
+  static GDisplayObject _findCommonParent(GDisplayObject obj1,
+      GDisplayObject obj2) {
     GDisplayObject? current = obj1;
 
     /// TODO: use faster Hash access.
@@ -776,9 +796,12 @@ abstract class GDisplayObject
   /// rendering.
   void requiresRedraw() {
     /// TODO: notify parent the current state of this DisplayObject.
-    $hasTouchableArea = visible && $maskee == null && _scaleX != 0 && _scaleY != 0;
+    $hasTouchableArea =
+        visible && $maskee == null && _scaleX != 0 && _scaleY != 0;
 
-    $hasVisibleArea = $alpha != 0 && visible && $maskee == null && _scaleX != 0 && _scaleY != 0;
+    $hasVisibleArea =
+        $alpha != 0 && visible && $maskee == null && _scaleX != 0 &&
+            _scaleY != 0;
   }
 
   void removedFromStage() {}
@@ -834,20 +857,23 @@ abstract class GDisplayObject
     final _hasTranslate = _x != 0 || _y != 0;
     final _hasPivot = _pivotX != 0 || _pivotY != 0;
     final _hasSkew = _skewX != 0 || _skewY != 0;
-    final needSave = _hasTranslate || _hasScale || rotation != 0 || _hasPivot || _hasSkew || _is3D;
+    final needSave = _hasTranslate || _hasScale || rotation != 0 || _hasPivot ||
+        _hasSkew || _is3D;
 
     final $hasFilters = hasFilters;
     // final hasColorize = $colorize?.alpha > 0 ?? false;
     // var _saveLayer = this is DisplayObjectContainer &&
     //     (this as DisplayObjectContainer).hasChildren &&
     //     ($alpha != 1 || $hasColorize || $hasFilters);
-    var _saveLayer = allowSaveLayer && $alpha != 1 || $hasColorize || $hasFilters;
+    var _saveLayer = allowSaveLayer && $alpha != 1 || $hasColorize ||
+        $hasFilters;
     // if (this is DisplayObjectContainer &&
     //     (this as DisplayObjectContainer).hasChildren) {
     // }
 
     final hasMask = mask != null || maskRect != null;
-    final showDebugBounds = DisplayBoundsDebugger.debugBoundsMode == DebugBoundsMode.internal &&
+    final showDebugBounds = DisplayBoundsDebugger.debugBoundsMode ==
+        DebugBoundsMode.internal &&
         ($debugBounds || DisplayBoundsDebugger.debugAll);
 
     GRect? _cacheLocalBoundsRect;
@@ -934,7 +960,8 @@ abstract class GDisplayObject
 
     if (_composerFilters != null) {
       for (var filter in _composerFilters) {
-        if (filter.hideObject || (filter is GDropShadowFilter && filter.innerShadow)) {
+        if (filter.hideObject ||
+            (filter is GDropShadowFilter && filter.innerShadow)) {
           filterHidesObject = true;
         }
         filter.process(canvas, $applyPaint);
@@ -998,8 +1025,9 @@ abstract class GDisplayObject
       ancestor = ancestor.$parent;
     }
     if (ancestor == this) {
-      throw ArgumentError('An object cannot be added as a child to itself or one '
-          'of its children (or children\'s children, etc.)');
+      throw ArgumentError(
+          'An object cannot be added as a child to itself or one '
+              'of its children (or children\'s children, etc.)');
     } else {
       $parent = value;
     }
@@ -1014,16 +1042,20 @@ abstract class GDisplayObject
     $setTransformationChanged();
   }
 
-  void setPosition(double x, double y) {
+  // shortcut to set x / y transformation
+  GDisplayObject setPosition(double x, double y) {
     _x = x;
     _y = y;
     $setTransformationChanged();
+    return this;
   }
 
-  void setScale(double scaleX, [double? scaleY]) {
+  // shortcut to set scaleX / scaleY transformation
+  GDisplayObject setScale(double scaleX, [double? scaleY]) {
     _scaleX = scaleX;
     _scaleY = scaleY ?? scaleX;
     $setTransformationChanged();
+    return this;
   }
 
   /// ---- capture texture feature ---
@@ -1032,7 +1064,8 @@ abstract class GDisplayObject
   /// transformations (x, y, scale, etc) if you intend to use in it's "original"
   /// form.
   ui.Picture createPicture(
-      [void Function(ui.Canvas)? prePaintCallback, void Function(ui.Canvas)? postPaintCallback]) {
+      [void Function(ui.Canvas)? prePaintCallback, void Function(ui
+          .Canvas)? postPaintCallback]) {
     final r = ui.PictureRecorder();
     final c = ui.Canvas(r);
     prePaintCallback?.call(c);
@@ -1063,7 +1096,8 @@ abstract class GDisplayObject
     if (resolution != 1) {
       rect *= resolution;
     }
-    final needsAdjust = (rect.left != 0 || rect.top != 0) && adjustOffset || resolution != 1;
+    final needsAdjust = (rect.left != 0 || rect.top != 0) && adjustOffset ||
+        resolution != 1;
     ui.Picture picture;
     if (needsAdjust) {
       picture = createPicture((canvas) {